feat: change short to int

+++ source/geometry/element.hpp
@@ -1,13 +1,13 @@
 class Element {
     private:
-        short ID;
+        int ID;
         Node* node1;
         Node* node2;
         Node* node3;
         Node* node4;
 
     public:
-        Element (short identifier, Node* first_node, Node* second_node, Node* third_node, Node* fourth_node){
+        Element (int identifier, Node* first_node, Node* second_node, Node* third_node, Node* fourth_node){
             ID = identifier;
             node1 = first_node;
             node2 = second_node;
@@ -15,10 +15,10 @@ class Element {
             node4 = fourth_node;
         }
 
-        void set_ID(short identifier){
+        void set_ID(int identifier){
             ID = identifier;
         }
-        short get_ID(){
+        int get_ID(){
             return ID;
         }
